Poly_commitment/utils: move pows into utils
